Remove alts, add Active Runner role, increase src.py scheduled job to 30min
Removing alts from the database to save on resources and to ensure the
integrity of the roles.

Adding Active Runner role, which will be given to runners with a run on
the leaderboard in the past 2 years.

Increasing to 30min as the jobs can overlap and overwrite each other,
which is not a good idea. Helps with rate limiting as well.